requirements.txt
Django==3.2.9
django-bootstrap3==21.1
django-filter==21.1
django-tables2==2.4.1
djangorestframework==3.12.4
psycopg2==2.9.1
setuptools==58.3.0
sqlparse==0.4.2
